Post-merge clean-ups
