+++ q2_moshpit/eggnog/__init__.py
@@ -9,20 +9,17 @@ from .orthologs import (
     eggnog_diamond_search, _eggnog_diamond_search, eggnog_hmmer_search,
     _eggnog_hmmer_search, _eggnog_feature_table,
 )
-from .annotation import eggnog_annotate, _eggnog_annotate
+from .annotation import eggnog_annotate, _eggnog_annotate, extract_annotations
 from .dbs import (
     fetch_eggnog_db, fetch_diamond_db, build_custom_diamond_db,
     fetch_eggnog_proteins, build_eggnog_diamond_db, fetch_ncbi_taxonomy,
     fetch_eggnog_hmmer_db
 )
-from .annotations import extract_annotations
-
 
 __all__ = [
     'eggnog_diamond_search', '_eggnog_diamond_search', 'eggnog_annotate',
     '_eggnog_feature_table', 'fetch_eggnog_db', 'fetch_diamond_db',
     'build_custom_diamond_db', 'fetch_eggnog_proteins',
     'build_eggnog_diamond_db', 'fetch_ncbi_taxonomy', '_eggnog_annotate',
-    'fetch_eggnog_hmmer_db', 'eggnog_hmmer_search', '_eggnog_hmmer_search',
-    'extract_annotations'
+    'fetch_eggnog_hmmer_db', 'eggnog_hmmer_search', '_eggnog_hmmer_search'
 ]
